Adding override switches
The overrides are to disable mechanisms that can harm the robot if
run unchecked.  Also, there is an override for the vision-based
odometry in the event a camera mount is busted or whatnot.

Also rewrote the LED code to match 6328's implementation... moved it
from a RBSI subsystem to a virtual subsystem.  Now, to cause changes
in the lights, write to one of the recognized boolean variables.  The
code checks all variables each loop and decorates the bot accordingly.

	modified:   src/main/java/frc/robot/Constants.java
	modified:   src/main/java/frc/robot/RobotContainer.java
	modified:   src/main/java/frc/robot/commands/ElevatorCommand.java
	deleted:    src/main/java/frc/robot/commands/LEDCommand.java
	modified:   src/main/java/frc/robot/subsystems/LED/LED.java
	deleted:    src/main/java/frc/robot/subsystems/LED/LEDIO.java
	deleted:    src/main/java/frc/robot/subsystems/LED/LEDIOCANdle.java
	modified:   src/main/java/frc/robot/subsystems/algae_mech/AlgaeMech.java
	modified:   src/main/java/frc/robot/subsystems/algae_mech/AlgaeMechIO.java
	modified:   src/main/java/frc/robot/subsystems/elevator/Elevator.java
	modified:   src/main/java/frc/robot/subsystems/elevator/ElevatorIO.java
	modified:   src/main/java/frc/robot/subsystems/elevator/ElevatorIOTalonFX.java
	modified:   src/main/java/frc/robot/subsystems/intake/Intake.java
	modified:   src/main/java/frc/robot/subsystems/vision/Vision.java
	modified:   src/main/java/frc/robot/util/OverrideSwitches.java